Bluetooth: handling AAC support for SPs not having VBR support
In case of incoming connection, DUT is not aware of the VBR capability of the remote. So, to expose AAC capability, we were adding remote in AAC WL.

In the fix, cached the VBR capability of the remote, which we use during   incoming connection.
During incoming connection, if the remote is VBR capable, and in not present in VBR BL, DUT shows AAC capability. If remote is present in VBR BL, we check in AAC Wl, if present, then DUT shows AAC capability.

CRs-fixed: 3982052


Change-Id: Iedf45f4b6bc67be952f8642b067f620cfa7517c0